database updates

+++ src/modules/auth/auth.service.ts
@@ -44,9 +44,15 @@ export class AuthService {
   }
 
   private async updateRefreshToken(refreshTokenDto: IRefreshToken) {
-    await this.prisma.sessions.delete({
+    const currentSession = await this.prisma.sessions.findFirst({
       where: { userId: refreshTokenDto.userId }
     });
+    if (currentSession) {
+      await this.prisma.sessions.delete({
+        where: { id: currentSession.id }
+      });
+    }
+
     return await this.prisma.sessions.create({
       data: refreshTokenDto
     });
